+++ src/CronJobs/ActivityLogAutoDeleteJob.php
@@ -36,16 +36,7 @@ class ActivityLogAutoDeleteJob implements ServiceInterface, ServiceCliInterface
 	 */
 	public function register(): void
 	{
-		if (!\apply_filters(SettingsFallback::FILTER_SETTINGS_GLOBAL_IS_VALID_NAME, false)) {
-			return;
-		}
-
-		if (!SettingsHelpers::isOptionCheckboxChecked(SettingsFallback::SETTINGS_FALLBACK_AUTO_DELETE_KEY, SettingsFallback::SETTINGS_FALLBACK_AUTO_DELETE_KEY)) {
-			return;
-		}
-
 		\add_action('admin_init', [$this, 'checkIfJobIsSet']);
-		\add_filter('cron_schedules', [$this, 'addJobToSchedule']); // phpcs:ignore WordPress.WP.CronInterval.ChangeDetected
 		\add_action(self::JOB_NAME, [$this, 'getJobCallback']);
 	}
 
@@ -59,27 +50,10 @@ class ActivityLogAutoDeleteJob implements ServiceInterface, ServiceCliInterface
 		if (!\wp_next_scheduled(self::JOB_NAME)) {
 			\wp_schedule_event(
 				\strtotime('tomorrow', \time()),
-				'daily',
+				CronJobsSchedules::CRON_JOBS_SCHEDULE_EVERY_DAY,
 				self::JOB_NAME
 			);
 		}
-	}
-
-	/**
-	 * Add job to schedule.
-	 *
-	 * @param array<mixed> $schedules WP schedules list.
-	 *
-	 * @return array<mixed>
-	 */
-	public function addJobToSchedule(array $schedules): array
-	{
-		$schedules['daily'] = [
-			'interval' => \DAY_IN_SECONDS,
-			'display' => \esc_html__('Every day at midnight', 'eightshift-forms'),
-		];
-
-		return $schedules;
 	}
 
 	/**
@@ -89,6 +63,14 @@ class ActivityLogAutoDeleteJob implements ServiceInterface, ServiceCliInterface
 	 */
 	public function getJobCallback()
 	{
+		if (!\apply_filters(SettingsFallback::FILTER_SETTINGS_GLOBAL_IS_VALID_NAME, false)) {
+			return;
+		}
+
+		if (!SettingsHelpers::isOptionCheckboxChecked(SettingsFallback::SETTINGS_FALLBACK_AUTO_DELETE_KEY, SettingsFallback::SETTINGS_FALLBACK_AUTO_DELETE_KEY)) {
+			return;
+		}
+
 		$logs = ActivityLogHelper::getActivityLogsAll(1, 100000);
 
 		$retentionInterval = SettingsHelpers::getOptionValueWithFallback(SettingsFallback::SETTINGS_FALLBACK_AUTO_DELETE_RETENTION_KEY, (string) SettingsFallback::SETTINGS_FALLBACK_AUTO_DELETE_RETENTION_DEFAULT_VALUE);